src/agents/customer-support-faq/requirements.txt
azure-ai-projects>=1.0.0
azure-ai-evaluation>=1.0.0
azure-identity>=1.17.0
openai>=1.50.0
pyyaml>=6.0.1
